splitting src/Arduino_ConnectionHandler.h into multiple files to avoid cyclic includes
